backend/pyproject.toml
[project]
name = "ai-playground-backend"
version = "0.1.0"
description = "Autonomous AI Data Scientist Playground - Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=2.0.0",
    "websockets>=14.0",
    "pydantic>=2.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
    "google-auth-httplib2>=0.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff",
]
